requirements.txt
pyusb
pyyaml
keybind
webcolors